requirements.txt
pip>=9
setuptools>=37
flask
flask_mqtt
pillow
pyserial
twisted
twisted-mqtt
treq
pyOpenSSL
service_identity
pyasn1
pyasn1-modules
bs4
python-dateutil
pytz
jsonpath-rw
astral
